requirements.txt
chardet==5.0.0
certifi==2022.6.15.1
click==8.1.3
firebase-admin==5.3.0
Flask==2.2.2
functions-framework==3.2.0
google-cloud-storage==2.5.0
google-cloud-secret-manager==2.12.4
google-auth==2.11.0
google-api-python-client==2.60.0
google-cloud-logging==3.2.2
idna==3.3
more-itertools==8.14.0
multiprocessing-logging==0.3.3
pathlib==1.0.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
retrying==1.3.3
six==1.16.0
structlog==22.1.0
urllib3==1.26.3              # We're trying 1.26.3 over 1.26.4 because of the change in the SSLContext default behavior. See release notes.
urllib3[secure]==1.26.3      #
